search-mcp/pyproject.toml
[project]
name = "search-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.2.0",
    "qdrant-client>=1.9.0",
    "llama-index-core>=0.10.0",
    "llama-index-vector-stores-qdrant>=0.2.0",
    "llama-index-embeddings-openai>=0.1.0",
    "fastembed>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
search-mcp = "server:main"

[tool.setuptools]
py-modules = ["server"]
